backend/pyproject.toml
[project]
name = "resumematch-backend"
version = "0.1.0"
description = "ResumeMatch API — lexical + semantic resume scoring"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "python-multipart==0.0.9",
    "pdfplumber==0.11.0",
    "python-docx==1.1.2",
    "scikit-learn==1.4.2",
    "numpy==1.26.4",
    "sentence-transformers==3.0.1",
    "python-dotenv==1.0.1",
    "pydantic==2.7.1",
]

[tool.uv]
dev-dependencies = []
